replacing the interceptor class provided by the Strut2-oval plugin. It lacks the BeanValidationAnnotationsConfigurer configuerer being added to the validator (therefore, by default the plugin's interceptor does not support JSR303 annotations).
Also the way the error message is returned has been modified to return an i18n key instead of a generic error string. With the i18n key we can easily map the error to our own custom string.

With this modifications, an external validation xml file is not needed anymore.