fix to bounding box filter reported by markus in a email:
Tim,
I tried to use this butterfly for the demo and then zoom into the china records like this:

http://uat.gbif.org/portal/occurrence/search?taxon_key=1894840&GEOMETRY=91.01074218749999%2020.96143961409685,91.01074218749999%2037.125286284966776,118.564453125%2037.125286284966776,118.564453125%2020.96143961409685,91.01074218749999%2020.96143961409685

But Ill get quite some records from Iran or turkey and the details Ive checked actually have coordinates from Iran. 
http://uat.gbif.org/portal/occurrence/185794032

So it seems the index does not do right?


Markus

the fix: seems that leaflet draw doesn't use the same ordering of coordinates for rectangles and bboxes