addresses concern from Tim H.:
One other thing I think is important for the public release. On the data publisher detail page e.g. http://uat.gbif.org/organization/a98e4f40-fa3b-11d8-a1a1-b8a03c50a862 we are still using the word Organization - both in the subtitle and Organization Information on the main tab. I realize this is the terminology that has been used in the registry but it is inconsistent with our references elsewhere (e.g. on homepage and all our communications). To avoid confusion we need to stick to one terminology, and I would strongly recommend Data publisher in both cases.